requirements.txt
numpy
scipy
pandas
quail
seaborn
matplotlib
sqlalchemy
nltk
scikit-learn
hypertools
pydata-wrangler
pyspellchecker
nltk
brainiak
flair
pycircstat
statsmodels
tqdm
